Add sass support and basic color variable
Added sass and sass-loader as project dependencies in package-lock.json. Created a new file named _variables.scss inside src/assets/sass to define basic color variables. The styles.css file is imported in main.ts to make the styles available globally. This is the initial step for setting up a design system for the app.