Add preview feature and refactor app structure
Introduced new Preview feature in the app and made a significant restructure of the Vue components and routings. Removed the Navbar component from App.vue and considered it under MainLayout. Created new 'preview' route and a navigation button for it in the Navbar. Changed global padding in reset.scss for better design. Also renamed 'logout-btn.vue' for better organization. These changes improve the maintainability and scalability of the project, and introduce a preview function for users.