"Adjust transition speed and add 'preview' button to navbar"
Reduced transition speed from 0.2s to 0.15s in mixins.scss file to create a faster and smoother user experience.

In addition, a 'preview' button was included in the navbar component to improve functionality. This button has defined styles for visibility and interactivity, utilizing the faster transition speed for a better user response.