Add drop-down component for platform selection
A new component 'drop-down.vue' was introduced to replace the select element used for platform selection in 'link.vue'. The component makes it possible to display icons and add more platforms easily. This changes increase flexibility and improves the UX. Removed font-size property as it was redundant due to global styles. A new interface 'dropdown-item.ts' also created to standardize the data structure.