Add transition effect and handle overflow in dropdown component
The "link" component has been updated to incorporate transition effect to the dropdown component. This creates a smoother user experience by adding an animation to the toggle behavior of the dropdown. The dropdown list now transitions in instead of instantly appearing. Additionally, a scrollbar has been added to the dropdown in case the list becomes too long. To accommodate the scrollbar, the dropdown's height and position have been adjusted. Furthermore, the click event on the input field has been moved to the parent div instead of the input element itself in order to avoid unintentional dropdown triggers. This change improves user interaction with the dropdown component. CSS styles related to transitions and the new scrollbar have been moved to their separate stylesheets for clearer code structure and reusability.