Refactor dropdown and integrate PrimeVue for UI components
This refactor includes the removal of `dropdown.vue` component and `dropdown.ts` store and the creation of the new `profile-link.ts` store. The UI for dropdown is now using PrimeVue's dropdown component to provide a better UX - no custom management is needed for open/close states of dropdown and it looks more standardized. A new interface `profile-link.ts` has also been introduced to better structure the data. This commit also enables the functionality to remove and add new profile links dynamically. This change enhances the maintainability and robustness of the code while providing more flexible and user-friendly interactions in the links sections.