Modify selectedPlatform ref and its usage in the link component
Changed the initial value of selectedPlatform reference in the link.vue component to have no initial value instead of an empty object. This is done due to unnecessary presence of initial empty object which was never supposed to hold any value in the first place. Also added optional chaining operator in selectedPlatform usage within 'contains-platform' test function as it may be undefined leading to lowerCase error.