Enable ripple effect in PrimeVue, refactor dropdown and scrollbar styles
This commit enables ripple effect in PrimeVue to provide a visual feedback when user interacts with the application. It helps enhance user experience by providing a visual confirmation of input.

The dropdown items styles are refractored, moving `max-height` styling from `.p-dropdown-items-wrapper` to `.p-dropdown-items`. This ensures the height limitation is applied to the actual dropdown items list.

Changes on scrollbar include limiting scrollbar width and thumb color styling to the body. This provides consistent styles to all scrollbars throughout the application.