Refactor form validation in link.vue and enhance UI elements
Form validation for platform selection and URL is separated and rewritten to provide more specific validation error messages. The validation is now focused on each field individually, which increases the readability of the validation code and allows the utilization of schema validation library(Yup) more effectively.

An icon of the chosen platform is also added to the UI dropdown menu for better user experience. Form inputs are now cleared after submission for better UX. A dynamic placeholder is added to the URL input field based on the platform selected.

Small changes in code formatting were made in main.ts according to our project style-guide.