Add unique id property to ProfileLink interface
Modified the ProfileLink interface by adding a unique 'id' property. This change was made to allow for stable tracking of each profile link within the app. Previously, the index was used as the key in Links.vue, but this could cause problems if the order of the items changes. Now, every ProfileLink object created by addProfileLink() function in the profile-link store will have a unique id which is used as a key in Links.vue.