"Enhance platform visual integration and optionalize ID
The commit introduces enhancements for platform visuals by adding a 'background' property for more accurate representation. A 'calculateYPosition' function was also integrated to improve SVG rendering in 'phone.vue'. An 'id' within 'ProfileLink' interface has been made optional. This is to allow more flexibility in instances where an 'id' might not be available or needed. Cleaning of logging in some components was also done."