Make ProfileLink fields optional and adjust button styling
The 'platform' and 'url' fields in the ProfileLink interface have been made optional to allow more flexibility when creating new links. Padding for the 'save-btn' in 'link.vue' component was also reduced for a more compact look. An 'align-items: center;' rule was added in 'phone.vue' for better element alignment in the flex container.