Add profile details form and remove console.log
Added a form to the profile view to allow users to input their information such as name, email and profile picture. This will enable a more personalized user experience. Also, removed a console.log from the link.vue file that was used for testing and is not needed in production.