Update conditions for displaying placeholder in phone component
This commit changes the conditions for the placeholder rectangle in the phone component. Previously, this rectangle was displayed if the user's firstName was not filled in. Now, it will only be displayed if neither the user's firstName nor lastName are filled in. This change makes the display of the placeholder more precise and accurate in the context