Add URL copy to clipboard functionality with toast notifications
In the Preview.vue view file, functionality was added to copy the current URL to the user's clipboard. This is achieved by using the native 'navigator.clipboard.writeText()' method. Toast notifications from the 'primevue/usetoast' package has been utilized to provide the user with feedback on whether the copying process was successful. This functionality is triggered by a button click event. The ToastService was also imported in the main.ts file to enable the use of toast notifications.