Adding Draggable Functionality with the vuedraggable Package
The vuedraggable package has been introduced as a part of the dependencies in order to facilitate drag and drop actions for reordering of profiles. The impact of this addition is observed in terms of improved user experience with an elevated level of customization in the profile section. The phone.vue component has been updated accordingly to actualize this feature.