Update variable naming for clarity and accuracy
Changed the term 'profileLinkIndex' to 'profileLinkId' in several areas of the code. This was done to better reflect the constant's purpose, and to prevent confusion as it is not an index but a unique id. This change impacts 'link.vue', 'profile-link.ts', and 'Links.vue' files. The handling of updating the profile link was also updated in 'profile-link.ts' to reflect this change, now finding the link based on the id.