Add placeholder for empty links list
The commit introduces a placeholder in the Links.vue component for when there are no profile links.
When there are no links to show, we now display a centered message with instructions on how to get
started, along with an illustration. This makes our application more user-friendly and helps
the user to understand what actions can be taken next.