Enable JS files compilation in TypeScript config
This commit modifies both `tsconfig.app.json` and `tsconfig.json` files by adding the "allowJs" property and setting it to true. This change enables JavaScript files to be compiled by the TypeScript compiler, including them in the build process. The reason behind this modification is to enable flexibility of having JS files in TypeScript project.